Use SpeciesConstructor; invoke .then; refactor closures
 - Use SpeciesConstructor to ensure subclasses are respected
 - Observably invoke `.then` with the result of the `onFinally` function
 - Refactor closures to use the more verbose form the spec currently favors.

Fixes #29. Fixes #32. Fixes #33.
